Move cluster components into the analysis folder
